calculo enderc

+++ AOCII/TI/TAD.c
@@ -12,8 +12,8 @@ typedef struct info
     int Capac_Conj;
     int Tam_Conj;
     int Num_Conjs;
-    int Tam_Overhead;
-    int Tam_MemTotal;
+    float Tam_Overhead;
+    float Tam_MemTotal;
 } INFO;
 
 typedef struct address
